+++ src/modules/downloader.py
@@ -2,14 +2,15 @@
 """
 This module interacts with the instagram api and uploads content to a temporary directory.
 Supports downloading the content of the post by link and getting information about the account.
-https://instaloader.github.io/module/instaloader.html
+https://github.com/subzeroid/instagrapi
 """
+import os
 from typing import Union
-from ast import literal_eval
-import base64
-import instaloader
+from pathlib import Path
+from instagrapi import Client
+from instagrapi.exceptions import LoginRequired
 from logger import log
-from .exceptions import WrongVaultInstance, FailedCreateDownloaderInstance, FailedAuthInstaloader
+from .exceptions import WrongVaultInstance, FailedCreateDownloaderInstance, FailedAuthInstagram
 
 
 # pylint: disable=too-few-public-methods
@@ -30,12 +31,15 @@ class Downloader:
             :param configuration (dict): dictionary with configuration parameters for Instagram API communication.
                 :param username (str): username for authentication in the instagram api.
                 :param password (str): password for authentication in the instagram api.
-                :param login-method (str): method for authentication in the instagram api. Can be: 'session', 'password', 'anonymous'.
-                :param session-file (str): the path for saving the session file.
-                :param user-agent (str): user-agent header.
-                :param fatal-status-codes (list): list of fatal status codes, this causes the thread executing this module's code to crash.
-                :param iphone-support (bool): enable or disable iphone http headers.
-                :param session-base64 (str): base64 encoded session file.
+                :param session-file (str): path to the session file for authentication in the instagram api.
+                :param delay-requests (int): delay between requests.
+                :param 2fa-enabled (bool): two-factor authentication enabled.
+                :param 2fa-seed (str): seed for two-factor authentication (secret key).
+                :param locale (str): locale for requests.
+                :param country-code (str): country code for requests.
+                :param timezone-offset (int): timezone offset for requests.
+                :param user-agent (str): user agent for requests.
+                :param proxy-dsn (str): proxy dsn for requests.
             :param vault (object): instance of vault for reading configuration downloader-api.
 
         Returns:
@@ -43,16 +47,21 @@ class Downloader:
 
         Attributes:
             :attribute configuration (dict): dictionary with configuration parameters for instagram api communication.
-            :attribute instaloader (object): instance of the instaloader class for working
+            :attribute client (object): instance of the instagram api client.
 
         Examples:
             >>> configuration = {
             ...     'username': 'my_username',
             ...     'password': 'my_password',
-            ...     'login-method': 'session',
-            ...     'session-file': '/path/to/session/file',
-            ...     'session-base64': '<base64_encoded_session_file>',
-            ...     'user-agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/58.0.3029.110 Safari/537.3'
+            ...     'session-file': 'data/session.json',
+            ...     'delay-requests': 1,
+            ...     '2fa-enabled': False,
+            ...     '2fa-seed': 'my_seed_secret',
+            ...     'locale': 'en_US',
+            ...     'country-code': '1',
+            ...     'timezone-offset': 10800,
+            ...     'user-agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64)',
+            ...     'proxy-dsn': 'http://localhost:8080'
             ... }
             >>> vault = Vault()
             >>> downloader = Downloader(configuration, vault)
@@ -69,30 +78,23 @@ class Downloader:
                 "Failed to initialize the Downloader instance."
                 "Please check the configuration in class argument or the secret with the configuration in the Vault."
             )
+
         log.info('[Downloader]: Creating a new instance...')
-        self.instaloader = instaloader.Instaloader(
-            quiet=True,
-            user_agent=self.configuration.get('user-agent', None),
-            iphone_support=self.configuration.get('iphone-support', None),
-            dirname_pattern='data/{profile}',
-            filename_pattern='{profile}_{shortcode}_{filename}',
-            download_pictures=True,
-            download_videos=True,
-            download_video_thumbnails=False,
-            download_geotags=False,
-            download_comments=False,
-            save_metadata=False,
-            compress_json=False,
-            post_metadata_txt_pattern=None,
-            storyitem_metadata_txt_pattern=None,
-            check_resume_bbd=True,
-            fatal_status_codes=literal_eval(self.configuration.get('fatal-status-codes', '[]'))
-        )
+        self.client = Client()
+
+        log.info('[Downloader]: Configuring client settings...')
+        self.client.delay_range = [1, int(self.configuration['delay-requests'])]
+        self.client.set_locale(locale=self.configuration['locale'])
+        self.client.set_country_code(country_code=int(self.configuration['country-code']))
+        self.client.set_timezone_offset(seconds=int(self.configuration['timezone-offset']))
+        self.client.set_user_agent(user_agent=self.configuration['user-agent'])
+        self.client.set_proxy(dsn=self.configuration.get('proxy-dsn', None))
+
         auth_status = self._login()
         if auth_status == 'logged_in':
             log.info('[Downloader]: Instance created successfully with account %s', self.configuration['username'])
         else:
-            raise FailedAuthInstaloader("Failed to authenticate the Instaloader instance.")
+            raise FailedAuthInstagram("Failed to authenticate the Instaloader instance.")
 
     def _login(self) -> Union[str, None]:
         """
@@ -105,33 +107,40 @@ class Downloader:
                 or
             None
         """
-        if self.configuration['login-method'] == 'session':
-            # If session-base64 defined in the configuration, then decode and save the session file.
-            if self.configuration.get('session-base64', None):
-                with open(self.configuration['session-file'], 'wb') as file:
-                    file.write(base64.b64decode(self.configuration['session-base64']))
-            # Otherwise, it expects the session file to be in the specified path.
-            self.instaloader.load_session_from_file(
-                self.configuration['username'],
-                self.configuration['session-file']
-            )
-            log.info('[Downloader]: Session file %s was saved successfully', self.configuration['session-file'])
-            return 'logged_in'
+        log.info('[Downloader]: Authentication in the Instagram API...')
 
-        if self.configuration['login-method'] == 'password':
-            self.instaloader.login(
-                self.configuration['username'],
-                self.configuration['password']
-            )
-            self.instaloader.save_session_to_file(self.configuration['session-file'])
-            log.info('[Downloader]: Login with password was successful. Saved session in %s', self.configuration['sessionfile'])
-            return 'logged_in'
+        if self.configuration['2fa-enabled']:
+            totp_code = self.client.totp_generate_code(seed=self.configuration['2fa-seed'])
+            log.info('[Downloader]: Two-factor authentication is enabled. TOTP code: %s', totp_code)
+            login_args = {
+                'username': self.configuration['username'],
+                'password': self.configuration['password'],
+                'verification_code': totp_code
+            }
+        else:
+            login_args = {
+                'username': self.configuration['username'],
+                'password': self.configuration['password']
+            }
 
-        if self.configuration['login-method'] == 'anonymous':
-            log.warning('[Downloader]: Initialization without authentication into an account (anonymous)')
-            return None
+        if os.path.exists(self.configuration['session-file']):
+            self.client.load_settings(self.configuration['session-file'])
+            self.client.login(**login_args)
+        else:
+            self.client.login(**login_args)
+            self.client.dump_settings(self.configuration['session-file'])
 
-        raise FailedAuthInstaloader("Failed to authenticate the Instaloader instance.")
+        try:
+            self.client.get_timeline_feed()
+        except LoginRequired:
+            log.error('[Downloader]: Authentication in the Instagram API failed.')
+            old_session = self.client.get_settings()
+            self.client.set_settings({})
+            self.client.set_uuids(old_session["uuids"])
+            self.client.login(**login_args)
+
+        log.info('[Downloader]: Authentication in the Instagram API was successful.')
+        return 'logged_in'
 
     def get_post_content(
         self,
@@ -153,25 +162,50 @@ class Downloader:
         """
         log.info('[Downloader]: Downloading the contents of the post %s...', shortcode)
         try:
-            post = instaloader.Post.from_shortcode(self.instaloader.context, shortcode)
-            self.instaloader.download_post(post, '')
-            log.info('[Downloader]: The contents of the post %s have been successfully downloaded', shortcode)
-            status = 'completed'
-            owner = post.owner_username
-            typename = post.typename
-        except instaloader.exceptions.BadResponseException as error:
-            log.error('[Downloader]: Error downloading post content: %s', error)
-            if "Fetching Post metadata failed" in str(error):
-                status = 'source_not_found'
-                owner = 'undefined'
-                typename = 'undefined'
-                log.warning('[Downloader]: Post %s not found, perhaps it was deleted. Message will be marked as processed.', shortcode)
-            else:
-                raise instaloader.exceptions.BadResponseException(error)
+            media_pk = self.client.media_pk_from_code(code=shortcode)
+            media_info = self.client.media_info(media_pk=media_pk).dict()
 
-        return {
-            'post': shortcode,
-            'owner': owner,
-            'type': typename,
-            'status': status
-        }
+            path = Path(f"data/{media_info['user']['username']}")
+            os.makedirs(path, exist_ok=True)
+
+            if media_info['media_type'] == 1:
+                self.client.photo_download(media_pk=media_info['pk'], folder=path)
+
+            elif media_info['media_type'] == 2 and media_info['product_type'] == 'feed':
+                self.client.video_download(media_pk=media_info['pk'], folder=path)
+
+            elif media_info['media_type'] == 2 and media_info['product_type'] == 'clips':
+                self.client.clip_download(media_pk=media_info['pk'], folder=path)
+
+            elif media_info['media_type'] == 8:
+                self.client.album_download(media_pk=media_info['pk'], folder=path)
+
+            else:
+                log.warning('[Downloader]: The media type is not supported for download: %s', media_info)
+
+            if os.listdir(path):
+                log.info('[Downloader]: The contents of the post %s have been successfully downloaded', shortcode)
+                response = {
+                    'post': shortcode,
+                    'owner': media_info['user']['username'],
+                    'type': {media_info['product_type'] if media_info['product_type'] else 'photo'},
+                    'status': 'completed'
+                }
+
+            else:
+                log.error('[Downloader]: Temporary directory is empty: %s', path)
+                response = {
+                    'post': shortcode,
+                    'owner': media_info['user']['username'],
+                    'type': {media_info['product_type'] if media_info['product_type'] else 'photo'},
+                    'status': 'failed'
+                }
+
+            return response
+
+        # pylint: disable=broad-except
+        # Temporary general exception for migration to the new module.
+        # Will be replaced by specific exceptions after v3.0.0
+        except Exception as error:
+            log.error('[Downloader]: Error downloading post content: %s\n%s', error, media_info)
+            return None