+++ src/modules/uploader.py
@@ -5,11 +5,9 @@ and uploads the found media files (image, video) to the destination storage.
 """
 import os
 from typing import Union
-import dropbox
-from mega import Mega
 from webdav3.client import Client as WebDavClient
 from logger import log
-from .exceptions import WrongVaultInstance, FailedInitUploaderInstance, WrongStorageType
+from .exceptions import WrongVaultInstance, FailedInitUploaderInstance
 
 
 class Uploader:
@@ -29,8 +27,6 @@ class Uploader:
             :param configuration (dict): dictionary with target storage parameters.
                 :param username (str): username for authentication in the target storage.
                 :param password (str): password for authentication in the target storage.
-                :param storage-type (str): type of storage for uploading content. Can be: 'dropbox', 'mega' or 'webdav'.
-                :param exclude-types (str): exclude files with this type from uploading. Example: '.json, .txt'.
                 :param source-directory (str): the path to the local directory with media content for uploading.
                 :param destination-directory (str): a subdirectory in the cloud storage where the content will be uploaded.
             :param vault (object): instance of vault for reading authorization data.
@@ -42,9 +38,7 @@ class Uploader:
             >>> configuration = {
             ...     'username': 'my_username',
             ...     'password': 'my_password',
-            ...     'storage-type': 'webdav',
             ...     'url': 'https://webdav.example.com/directory',
-            ...     'exclude-types': '.json, .txt',
             ...     'source-directory': '/path/to/source/directory',
             ...     'destination-directory': '/path/to/destination/directory'
             ... }
@@ -64,37 +58,17 @@ class Uploader:
                 "Please check the configuration in class argument or the secret with the configuration in the Vault."
             )
 
+        log.info('[Uploader]: Initializing connection to the WebDav remote directory...')
         self.local_directory = f"{os.getcwd()}/{self.configuration['source-directory']}"
-        self.storage = self._init_storage_connection()
+        options = {
+            'webdav_hostname': self.configuration['url'],
+            'webdav_login': self.configuration['username'],
+            'webdav_password': self.configuration['password']
+        }
+        self.storage = WebDavClient(options)
+        log.info('[Uploader]: Connection to the WebDav remote directory is established')
+
         self._check_incomplete_transfers()
-
-    def _init_storage_connection(self) -> object:
-        """
-        The method for initializing a connection to the target storage.
-        Can be: 'dropbox', 'mega' or 'webdav'.
-
-        Args:
-            None
-
-        Returns:
-            (object) connection object to the target storage.
-        """
-        if self.configuration['storage-type'] == 'dropbox':
-            return dropbox.Dropbox(oauth2_access_token=self.configuration['password'])
-
-        if self.configuration['storage-type'] == 'mega':
-            mega = Mega()
-            return mega.login(email=self.configuration['username'], password=self.configuration['password'])
-
-        if self.configuration['storage-type'] == 'webdav':
-            options = {
-                'webdav_hostname': self.configuration['url'],
-                'webdav_login': self.configuration['username'],
-                'webdav_password': self.configuration['password']
-            }
-            return WebDavClient(options)
-
-        raise WrongStorageType("Wrong storage type, please check the configuration. 'storage-type' can be: 'dropbox', 'mega' or 'webdav'.")
 
     def _check_incomplete_transfers(self) -> None:
         """
@@ -137,21 +111,18 @@ class Uploader:
         """
         transfers = {}
         result = ""
-        log.info('[Uploader]: Preparing media files for transfer to the %s...', self.configuration['storage-type'])
+        log.info('[Uploader]: Preparing media files for transfer to the cloud...')
         for root, _, files in os.walk(f"{self.configuration['source-directory']}{sub_directory}"):
             for file in files:
-                if file.split('.')[-1] in self.configuration.get('exclude-types', None):
+                transfers[file] = self.upload_to_cloud(
+                    source=os.path.join(root, file),
+                    destination=root.split('/')[1]
+                )
+                if transfers[file] == 'uploaded':
                     os.remove(os.path.join(root, file))
+                    result = 'completed'
                 else:
-                    transfers[file] = self.upload_to_cloud(
-                        source=os.path.join(root, file),
-                        destination=root.split('/')[1]
-                    )
-                    if transfers[file] == 'uploaded':
-                        os.remove(os.path.join(root, file))
-                        result = 'completed'
-                    else:
-                        result = 'not_completed'
+                    result = 'not_completed'
         log.info('[Uploader]: List of all transfers %s', transfers)
         return result
 
@@ -172,38 +143,18 @@ class Uploader:
                 or
             None
         """
-        log.info('[Uploader]: Starting upload file %s to %s://%s', source, self.configuration['storage-type'], destination)
-        response = None
-        result = None
+        log.info('[Uploader]: Starting upload file %s to WebDav://%s', source, destination)
 
-        if self.configuration['storage-type'] == 'mega':
-            directory = f"{self.configuration['destination-directory']}/{destination}"
-            log.info('[Uploader]: Trying found mega folder %s...', directory)
-            mega_folder = self.storage.find(directory, exclude_deleted=True)
-            if not mega_folder:
-                self.storage.create_folder(directory)
-                mega_folder = self.storage.find(directory, exclude_deleted=True)
-                log.info('[Uploader]: Mega folder not found, created new folder %s', mega_folder)
-            else:
-                log.info('[Uploader]: Mega folder %s was found', mega_folder)
-            response = self.storage.upload(filename=source, dest=mega_folder[0])
-            result = "uploaded"
+        if not self.storage.check(f"{self.configuration['destination-directory']}/{destination}"):
+            self.storage.mkdir(f"{self.configuration['destination-directory']}/{destination}")
+        self.storage.upload_sync(
+            remote_path=f"{self.configuration['destination-directory']}/{destination}/{source.split('/')[-1]}",
+            local_path=source
+        )
 
-        if self.configuration['storage-type'] == 'dropbox':
-            with open(source, 'rb') as file_transfer:
-                response = self.storage.files_upload(file_transfer.read(), f"/{destination}/{source.split('/')[-1]}")
-            file_transfer.close()
-            result = "uploaded"
-
-        if self.configuration['storage-type'] == 'webdav':
-            if not self.storage.check(f"{self.configuration['destination-directory']}/{destination}"):
-                self.storage.mkdir(f"{self.configuration['destination-directory']}/{destination}")
-            self.storage.upload_sync(
-                remote_path=f"{self.configuration['destination-directory']}/{destination}/{source.split('/')[-1]}",
-                local_path=source
-            )
-            response = self.storage.info(f"{self.configuration['destination-directory']}/{destination}/{source.split('/')[-1]}")['etag']
-            result = "uploaded"
-
-        log.info('[Uploader]: %s successful transferred', response)
-        return result
+        status = self.storage.info(f"{self.configuration['destination-directory']}/{destination}/{source.split('/')[-1]}")
+        if status['etag']:
+            log.info('[Uploader]: %s successful transferred in WebDav directory', status['etag'])
+            return "uploaded"
+        log.error('[Uploader]: failed to transfer in WebDav directory: %s', source)
+        return None